// src/enums/Reference.ts
// ----------------------------------------------------------------
// PURPOSE:
// This library implements the Enum for supported References.
// ----------------------------------------------------------------

export enum Reference {
    FONTS = '@fonts',
    IMAGES = '@image',
    BOOKS = '@books',
    CONFIG = '@config',
    PERSONS = '@person',
    CHAPTERS = '@chapters',
    STORIES = '@stories',
    CODE = '@code',
    LEGAL = '@legal',
    TOPICS = '@topic',
    FILES = '@file',
    I18N = '@i18n',
}